+++ src/component/sources/cache.ts
@@ -2,7 +2,7 @@ import { TagifyResponseItem } from '../../client/TagifyClient';
 
 const LIMIT_KEY: string = 'tagify-limit';
 const API_VERSION_KEY: string = 'tagify-api-version';
-const DEFAULT_CACHE_TTL: number = 1000 * 60 * 60 * 24 * 30; // 30 days
+const DEFAULT_CACHE_TTL: number = 1000 * 60 * 60 * 24 * 90; // 90 days
 const DEFAULT_ASYNC_CACHE_TTL: number = 1000 * 60 * 15; // 15 minutes
 
 export type CachedValue = TagLimit | ApiVersion | TagifyResponseItem;